Project Completed

+++ controller/places-controller.js
@@ -1,6 +1,8 @@
 const { validationResult } = require("express-validator");
 const HttpError = require("../models/http-error");
 const { v4: uuidv4 } = require("uuid");
+const fs = require("fs");
+
 const getCoordsForAddress = require("../util/location");
 const Place = require("../models/place");
 const User = require("../models/user");
@@ -101,7 +103,7 @@ const createPlace = async (req, res, next) => {
       new HttpError("Invalid input passes, please check your data", 422)
     );
   }
-  const { title, description, address, creatorId } = req.body; // title = req.body.title
+  const { title, description, address } = req.body; // title = req.body.title
 
   try {
     location = getCoordsForAddress(address);
@@ -118,17 +120,20 @@ const createPlace = async (req, res, next) => {
     description,
     address,
     location,
-    image: "https://fastread.in/images/uploads/Galteshwar-Mahadev-Temple.jpg ",
-    creatorId,
+    image: req.file.path,
+    creatorId:req.userData.userId,
   });
   let user;
-  console.log(creatorId);
+  // console.log(creatorId);
   try {
-    user = await User.findById(creatorId).exec();
+    user = await User.findById(req.userData.userId);
   } catch (error) {
-    console.log(error)
+    console.log(error);
     return next(
-      new HttpError("Creating place failed (user not found), Please try again letter", 500)
+      new HttpError(
+        "Creating place failed (user not found), Please try again letter",
+        500
+      )
     );
   }
   if (!user) {
@@ -155,24 +160,17 @@ const createPlace = async (req, res, next) => {
 const updatePlace = async (req, res, next) => {
   const errors = validationResult(req);
   if (!errors.isEmpty()) {
+    console.log("error from update place", errors);
     return next(
       new HttpError("Invalid input passes, please check your data", 422)
     );
   }
-  const { title, description, address } = req.body; // title = req.body.title
+  const { title, description } = req.body; // title = req.body.title
   const placeId = req.params.pid;
 
-  try {
-    const updatedPlace = await Place.findOneAndUpdate(
-      { _id: placeId },
-      { title, description, address },
-      { new: true } // To return the updated document
-    );
-  } catch (e) {
-    return next(
-      new HttpError("Something went wrong, could not update place", 500)
-    );
-  }
+  console.log("from Update Place");
+  console.log(req.body);
+
   let place;
   try {
     place = await Place.findById(placeId);
@@ -182,39 +180,69 @@ const updatePlace = async (req, res, next) => {
       new HttpError("Something went wrong could not find the place", 500)
     );
   }
-  res.status(200).json({ place: place.toObject({ getters: true }) });
+  if (place.creatorId.toString() !== req.userData.userId) {
+    return next(new HttpError("You are not allowed to edit this place", 401));
+  }
+  let updatedPlace;
+
+  try {
+    updatedPlace = await Place.findOneAndUpdate(
+      { _id: placeId },
+      { title, description },
+      { new: true } // To return the updated document
+    );
+  } catch (e) {
+    return next(
+      new HttpError("Something went wrong, could not update place", 500)
+    );
+  }
+
+  console.log(updatePlace);
+  let newPlace;
+  try {
+    newPlace = await Place.findById(placeId);
+  } catch (error) {
+    console.log(error);
+    return next(
+      new HttpError("Something went wrong could not find the place", 500)
+    );
+  }
+  res.status(200).json({ place: newPlace.toObject({ getters: true }) });
 };
 const deletePlace = async (req, res, next) => {
   const placeId = req.params.pid;
+
   let place;
   try {
-    place = await Place.findById(placeId).populate('creatorId');
+    place = await Place.findById(placeId).populate("creatorId");
   } catch (error) {
-    console.log(error)
+    console.log(error);
     return next(
       new HttpError("Something went wrong could find the place", 500)
     );
   }
-  if(!place){
-    return next(
-      new HttpError("Could not find place for this ID", 404)
-    );
+  if (!place) {
+    return next(new HttpError("Could not find place for this ID", 404));
   }
+  // console.log(place.creatorId.id.toString()+"hello above is creator\n"+req.userData.userId)
+  if (place.creatorId.id.toString() !== req.userData.userId) {
+    return next(new HttpError("You are not allowed to edit this place", 401));
+  }
+  const imagePath = place.image;
   try {
     const session = await mongoose.startSession();
     await session.startTransaction();
     await Place.findByIdAndRemove(placeId, session);
     place.creatorId.places.pull(place);
-    await place.creatorId.save({session});
+    await place.creatorId.save({ session });
     await session.commitTransaction();
-
   } catch (error) {
-      console.log(error);
-      return next(
-        new HttpError("Something wrong during deletion", 500)
-      );
+    console.log(error);
+    return next(new HttpError("Something wrong during deletion", 500));
   }
-  
+  fs.unlink(imagePath, (err) => {
+    console.log(err);
+  });
   res.status(200).json({ message: "Sucessfully message got deleted" });
 };
